+++ api-processor/src/test/kotlin/io/bkbn/lerasium/api/processor/KtorProcessorProviderTest.kt
@@ -3,13 +3,14 @@ package io.bkbn.lerasium.api.processor
 import com.tschuchort.compiletesting.KotlinCompilation
 import com.tschuchort.compiletesting.SourceFile
 import com.tschuchort.compiletesting.symbolProcessorProviders
+import io.bkbn.lerasium.utils.TestUtils.kotlinCode
+import io.bkbn.lerasium.utils.TestUtils.kspGeneratedSources
+import io.bkbn.lerasium.utils.TestUtils.readTrimmed
 import io.kotest.core.spec.style.DescribeSpec
 import io.kotest.matchers.collections.shouldHaveSize
 import io.kotest.matchers.shouldBe
 import io.kotest.matchers.shouldNotBe
 import io.kotest.matchers.string.shouldInclude
-import org.intellij.lang.annotations.Language
-import java.io.File
 
 class KtorProcessorProviderTest : DescribeSpec({
   describe("Validation") {
@@ -245,22 +246,6 @@ class KtorProcessorProviderTest : DescribeSpec({
           interface UserApiSpec : UserDomain
         """.trimIndent()
     )
-
-    private val KotlinCompilation.Result.workingDir: File
-      get() =
-        outputDirectory.parentFile!!
-
-    val KotlinCompilation.Result.kspGeneratedSources: List<File>
-      get() {
-        val kspWorkingDir = workingDir.resolve("ksp")
-        val kspGeneratedDir = kspWorkingDir.resolve("sources")
-        val kotlinGeneratedDir = kspGeneratedDir.resolve("kotlin")
-        return kotlinGeneratedDir.walkTopDown().toList().filter { it.isFile }
-      }
-
-    fun File.readTrimmed() = readText().trim()
-
-    fun kotlinCode(@Language("kotlin") contents: String): String = contents
   }
 }