+++ rdbms-processor/src/test/kotlin/io/bkbn/lerasium/rdbms/processor/RdbmsProcessorProviderTest.kt
@@ -3,13 +3,15 @@ package io.bkbn.lerasium.rdbms.processor
 import com.tschuchort.compiletesting.KotlinCompilation
 import com.tschuchort.compiletesting.SourceFile
 import com.tschuchort.compiletesting.symbolProcessorProviders
+import io.bkbn.lerasium.utils.TestUtils.errorMessage
+import io.bkbn.lerasium.utils.TestUtils.kotlinCode
+import io.bkbn.lerasium.utils.TestUtils.kspGeneratedSources
+import io.bkbn.lerasium.utils.TestUtils.readTrimmed
 import io.kotest.core.spec.style.DescribeSpec
 import io.kotest.matchers.collections.shouldContain
 import io.kotest.matchers.collections.shouldHaveSize
 import io.kotest.matchers.shouldBe
 import io.kotest.matchers.shouldNotBe
-import org.intellij.lang.annotations.Language
-import java.io.File
 
 class RdbmsProcessorProviderTest : DescribeSpec({
   describe("Table Generation") {
@@ -849,26 +851,4 @@ class RdbmsProcessorProviderTest : DescribeSpec({
       fileNames shouldContain "WordsTable.kt"
     }
   }
-}) {
-  companion object {
-    const val errorMessage = "\"\"Unable to get entity with id: \$id\"\""
-    private val KotlinCompilation.Result.workingDir: File
-      get() =
-        outputDirectory.parentFile!!
-
-    val KotlinCompilation.Result.kspGeneratedSources: List<File>
-      get() {
-        val kspWorkingDir = workingDir.resolve("ksp")
-        val kspGeneratedDir = kspWorkingDir.resolve("sources")
-        val kotlinGeneratedDir = kspGeneratedDir.resolve("kotlin")
-        return kotlinGeneratedDir.walkTopDown().toList().filter { it.isFile }
-      }
-
-    fun File.readTrimmed() = readText().trim()
-
-    fun kotlinCode(
-      @Language("kotlin") contents: String,
-      postProcess: (String) -> String = { it }
-    ): String = postProcess(contents)
-  }
-}
+})